% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fun.R
\docType{data}
\name{COL_C_BASES}
\alias{COL_C_BASES}
\title{Default color scheme for modified cytosine nucleobases}
\format{
An object of class \code{character} of length 5.
}
\usage{
COL_C_BASES
}
\description{
Based on Brewer's "Paired" palette,
\code{RColorBrewer::brewer.pal(12, "Paired")}.
}
\keyword{datasets}
