% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fun.R
\docType{data}
\name{CPG_COMBINATIONS}
\alias{CPG_COMBINATIONS}
\title{CpG combinations}
\format{
An object of class \code{character} of length 25.
}
\usage{
CPG_COMBINATIONS
}
\description{
CpG combinations
}
\keyword{datasets}
