% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerr.R
\name{get_template}
\alias{get_template}
\title{Script from template}
\usage{
get_template(package = NULL, filename = "template", version = "")
}
\arguments{
\item{package}{The name of the corresponding package this template is distributed with.}

\item{filename}{A generic filename of the template, e.g., "fun_template".}

\item{version}{A version idtenfier, e.g., "A01".}
}
\description{
Opens a new file with a template.
}
