% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggers.R
\name{log_debugging}
\alias{log_debugging}
\title{Logging for debugging}
\usage{
log_debugging(..., object = NULL)
}
\arguments{
\item{...}{message}

\item{object}{An (optional) object that should be printed to the console.}
}
\description{
Will print the function name it was invoked from.
}
