% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_loggers.R
\name{log_error}
\alias{log_error}
\title{Handling of errors}
\usage{
log_error(header, body, ...)
}
\arguments{
\item{header}{A concise description of the error.}

\item{body}{A named character that indicates, e.g. which elements caused an
error.}

\item{...}{Anything else passed to \code{\link[rlang:abort]{abort(...)}} as metadata.}
}
\description{
Handling of errors
}
