% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fun.R
\docType{data}
\name{MOD_C_BASES}
\alias{MOD_C_BASES}
\title{Modified cytosine nucleobases}
\format{
An object of class \code{factor} of length 5.
}
\usage{
MOD_C_BASES
}
\description{
Modified cytosine nucleobases
}
\keyword{datasets}
