% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_fun.R
\docType{data}
\name{MODIFIED_C_BASES}
\alias{MODIFIED_C_BASES}
\title{Modified cytosine nucleobases}
\format{
An object of class \code{factor} of length 5.
}
\usage{
MODIFIED_C_BASES
}
\description{
Modified cytosine nucleobases
}
\keyword{datasets}
