% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_fun.R
\name{normalizePath}
\alias{normalizePath}
\title{Express file paths in their canonical form -- Truly}
\usage{
normalizePath(path, winslash = .Platform$file.sep, mustWork = FALSE)
}
\arguments{
\item{path}{A character vector of file paths.}

\item{winslash}{the separator to be used on Windows -- ignored
    elsewhere.  Must be one of \code{c("/", "\\\\")}.}

\item{mustWork}{logical: if \code{TRUE} then an error is given if the result
    cannot be determined; if \code{NA} then a warning.}
}
\value{
A character vector.
}
\description{
Under Windows and UNIX, \code{\link[base:.Platform]{.Platform$file.sep}} both
may expand to "\code{/}", whereas some upstream applications may still use
"\code{\\}" to separate folders and files.
}
\details{
Calls \code{\link[base:normalizePath]{normalizePath}} with \code{.Platform$file.sep}.
In addition, trailing slashes are removed.
}
