% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerr.R
\name{update_summerr}
\alias{update_summerr}
\title{Update the summerr suite}
\usage{
update_summerr(pkgs = NULL, git_repo = "benjbuch")
}
\arguments{
\item{pkgs}{The package to update. If \code{NULL}, all currently loaded \code{summerr}
packages will be updated.}

\item{git_repo}{A GitHub repository to fetch the latest versions from. If \code{NULL},
CRAN is tried.}
}
\description{
Update the summerr suite
}
