% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summerr.R
\docType{package}
\name{summerr-package}
\alias{summerr-package}
\title{summerr}
\description{
General helpers for lab fun.
}
\details{
summerr-package
}
\section{General Functions}{


\subsection{File and Directory Operations}{

\link{select_directory}

\link{select_single_file}

\link{backup_file}

}

\subsection{Importing Experiment Layouts}{

\link{import_layout_from_excel}

\link{import_layout_from_paths}

\link{display_plate_layout}

}

\subsection{Formatting Helpers}{

\link{as_well}

}
}

\author{
Benjamin Buchmuller \email{benjamin.buchmuller@tu-dortmund.de}
}
